Organized subcomponents for Product Component into a folder and handled file dependencies for that change
